# docs/swagger.yaml
openapi: 3.0.2
info:
    title: Malle-Andre-pank
    version: 1.0.0
servers:
    -
        url: 'http://localhost:1234'
        description: ''
paths:
    /users:
        post:
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/postUsersRequest'
                required: true
            tags:
                - users
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/userObject'
                    description: User successfully registred.
                '400':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/errorResponse'
                            examples:
                                Invalid password:
                                    value:
                                        error: Invalid password
                                Parameter is required:
                                    value:
                                        error: User validation failed
                    description: Bad request.
                '409':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/errorResponse'
                            examples:
                                Username already exists:
                                    value:
                                        error: Username already exists
                    description: Username already exists.
            summary: Registers a new user
    /sessions:
        post:
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/postSessionsRequest'
                required: true
            tags:
                - sessions
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/postSessionsResponse'
                    description: User logged in succesfully
                '400':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/errorResponse'
                            examples:
                                Missing username or password:
                                    value:
                                        error: Missing username or password
                    description: Missing username or password
                '401':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/errorResponse'
                            examples:
                                Invalid username or password:
                                    value:
                                        error: Invalid username or password
                    description: Invalid username or password
            summary: Logs the user in
        delete:
            tags:
                - sessions
            responses:
                '204':
                    description: User logged out successfully
                '401':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/errorResponse'
                            examples:
                                Missing Authorization header:
                                    value:
                                        error: Missing Authorization header
                                Invalid token:
                                    value:
                                        error: Invalid token
                    description: Unauthorized
            summary: Logs the user out
    /users/current:
        get:
            tags:
                - users
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/getUsersCurrentResponse'
                    description: Successful request
            security:
                -
                    Token: []
            summary: Returns currently logged in user's data
    /transactions:
        get:
            tags:
                - transactions
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/getTransactionResponse'
                    description: Returns an array of user's transactions
                '401':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/errorResponse'
                            examples:
                                invalidToken:
                                    value:
                                        error: Invalid token
                    description: Invalid token
            security:
                -
                    Token: []
        post:
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/postTransactionsRequest'
                required: true
            tags:
                - transactions
            responses:
                '201':
                    content:
                        application/json: {}
                    description: Transactions has been successfully queued for transaction.
                '400':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/errorResponse'
                            examples:
                                invalidAmount:
                                    value:
                                        error: Invalid amount
                                invalidAccountTo:
                                    value:
                                        error: Invalid accountTo
                    description: Invalid amount
                '402':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/errorResponse'
                            examples:
                                paymentRequired:
                                    value:
                                        error: Payment required
                    description: Insufficient funds
                '403':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/errorResponse'
                            examples:
                                forbidden:
                                    value:
                                        error: Forbidden
                    description: AccountFrom does not belong to the user
                '404':
                    content:
                        application/json: {}
                    description: Account not found
            security:
                -
                    Token: []
    /transactions/b2b:
        post:
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/postTransactionsReceivingRequest'
                required: true
            tags:
                - transactions
            responses:
                '400':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/errorResponse'
                            examples:
                                JWTParsingFailed:
                                    value:
                                        error: Parsing JWT payload failed
                                invalidSignature:
                                    value:
                                        error: Invalid signature
                    description: Bad request
                '404':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/errorResponse'
                            examples:
                                accountNotFound:
                                    value:
                                        error: Account not found
                                accountNotInCentralBank:
                                    value:
                                        error: >-
                                            The account sending the funds does not belong to a bank registered
                                            in Central Bank
                    description: Account not found
                '500':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/errorResponse'
                            examples:
                                missingJwksUrl:
                                    value:
                                        error: 'Cannot verify your signature: The jwksUrl of your bank is missing'
                                invalidJwksUrl:
                                    value:
                                        error: 'Cannot verify your signature: The jwksUrl of your bank is invalid'
                    description: Cannot verify signature
                '502':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/errorResponse'
                            examples:
                                centralBankError:
                                    value:
                                        error: Central Bank error
                    description: Central Bank error
            security:
                -
                    Token: []
components:
    schemas:
        postUsersRequest:
            title: POST /users request
            description: ''
            required:
                - name
                - username
                - password
            type: object
            properties:
                name:
                    description: User's full name.
                    type: string
                username:
                    description: Chosen username.
                    type: string
                password:
                    format: password
                    description: User's password.
                    type: string
            example:
                name: Juku
                username: Uku
                password: wordpass
        userObject:
            title: user object
            description: ''
            type: object
            properties:
                id:
                    description: User's ID.
                    type: string
                name:
                    description: User's full name.
                    type: string
                username:
                    description: Chosen username'
                    type: string
                accounts:
                    description: Array of user's bank accounts.
                    type: array
                    items:
                        $ref: '#/components/schemas/account'
            example:
                id: 5f7b7ad24abdf90a746edaac
                name: Juku
                username: Uku
                accounts:
                    -
                        name: Main
                        balance: 100000
                        currency: EUR
                        number: abc84dbea2a7b4b39fac9cbb2d6469ebfab
                        id: 5f7b7ad34abdf90a746edaad
        errorResponse:
            title: error object
            description: ''
            type: object
            properties:
                error:
                    type: string
            example:
                error: Error message.
        postSessionsRequest:
            title: POST /sessions request
            description: ''
            type: object
            properties:
                username:
                    description: A valid username
                    type: string
                password:
                    description: A valid password
                    type: string
            example:
                username: Uku
                password: wordpass
        postSessionsResponse:
            title: POST /sessions response
            description: ''
            type: object
            properties:
                token:
                    description: A session token which needs to be send back with every request to the server
                    type: string
            example:
                token: 5f8022d6f40ea213f83577fc
        getUsersCurrentResponse:
            title: GET /users/current response
            description: ''
            type: object
            properties:
                id:
                    type: string
                name:
                    type: string
                username:
                    type: string
                accounts:
                    description: Array of user's accounts
                    type: array
                    items:
                        $ref: '#/components/schemas/account'
            example:
                id: 5f8015346fea363d746f935b
                name: Juku
                username: Uku
                accounts:
                    -
                        name: Main
                        balance: 100000
                        currency: EUR
                        number: abc0d3949b73af69489183f5e5d2760537c
        account:
            title: account object
            description: ''
            type: object
            properties:
                name:
                    type: string
                balance:
                    type: integer
                currency:
                    type: string
                number:
                    type: string
            example:
                name: Main
                balance: 100000
                currency: EUR
                number: abc0d3949b73af69489183f5e5d2760537c
        postTransactionsRequest:
            title: POST /transactions request
            description: ''
            type: object
            properties:
                accountFrom:
                    description: Sender's bank account
                    type: string
                accountTo:
                    description: Reciever's bank account
                    type: string
                amount:
                    format: int32
                    description: Transfered sum
                    type: integer
                explanation:
                    description: Purpose of the transaction
                    type: string
            example:
                accountFrom: RU000000001
                accountTo: BY000000001
                amount: 1500000000
                explanation: Don’t go spend it all at once
        postTransactionsReceivingRequest:
            title: POST /transactionsReceiving request
            description: ''
            type: object
            properties:
                accountFrom:
                    description: Sender's account
                    type: string
                accountTo:
                    description: Receiver's account
                    type: string
                amount:
                    format: int32
                    description: Transferred sum
                    type: integer
                currency:
                    description: Transaction's currency
                    type: string
                explanation:
                    description: Short description of transaction
                    type: string
                senderName:
                    description: Sender's name
                    type: string
            example:
                accountFrom: RU000000001
                accountTo: BY000000002
                amount: 1500000000
                currency: USD
                explanation: Thx buddy
                senderName: Aleksander Lukashenko
        getTransactionResponse:
            title: GET /transaction response
            description: ''
            required: []
            type: object
            properties:
                id:
                    description: Transaction ID
                    type: string
                status:
                    description: '''One of: ''''pending'''', ''''completed'''', ''''inProgress'''', ''''failed'''''''
                    type: string
                amount:
                    description: ''
                    type: string
                currency:
                    description: ''
                    type: string
                accountFrom:
                    description: Sender's account
                    type: string
                accountTo:
                    description: Receiver's account
                    type: string
                explanation:
                    description: ''
                    type: string
                senderName:
                    description: ''
                    type: string
                createdAt:
                    format: date-time
                    description: ''
                    type: string
                statusDetail:
                    description: ''
                    type: string
            example:
                id: 5fa0428ccfe1f221b8dd6db4
                status: inProgress
                userId: 5f8331c99cc259057459e29f
                amount: '3'
                currency: EUR
                accountFrom: abc3b770e63e9f1c2efc2b83c187653d2cc
                accountTo: BY000000001
                explanation: Don’t go spend it all at once
                statusDetail: ''
                senderName: Juku
                createdAt: '2020-11-02T17:31:56.780+00:00'
    securitySchemes:
        Token:
            scheme: bearer
            type: http
security:
    -
        Token: []